storage backend: Add RBD (RADOS Block Device) support
This patch adds support for a new storage backend with RBD support.

RBD is the RADOS Block Device and is part of the Ceph distributed storage system.

It comes in two flavours: Qemu-RBD and Kernel RBD, this storage backend only supports
Qemu-RBD, thus limiting the use of this storage driver to Qemu only.

To function this backend relies on librbd and librados being present on the local system.

The backend also supports Cephx authentication for safe authentication with the Ceph cluster.

For storing credentials it uses the build-in secret mechanism of libvirt.

Signed-off-by: Wido den Hollander <wido@widodh.nl>