rbd: Use librbd async I/O when possible during pool refresh
Librbd supports Asynchronous opening of RBD images when
LIBRBD_SUPPORTS_AIO_OPEN is defined.

The old implementation would iterate over all the RBD images in
a pool and call rbd_open() and rbd_stat() on each of them in a loop.

With thousands of images in a pool this could take minutes or even longer
depending on the latency to the Ceph cluster.

By opening the RBD image async we can use a callback to populate libvirt's
volume list and reduce the amount of time it takes to refresh a RBD pool.